# api/swagger.yaml
---
swagger: "2.0"
info:
  description: "The endpoint connects with database to provide GET and POST service."
  version: "1.0.0"
  title: "License to Purge API"
basePath: "/license-to-purge/v1"
tags:
- name: "Software Master List"
  description: "Provide CRUD operation on software master list."
- name: "User List"
  description: "Get user list and user software list based on admin"
paths:
  /getMasterList:
    get:
      tags:
      - "Software Master List"
      summary: "Get software list"
      description: "Gets entire monitored software list."
      operationId: "getSoftwareList"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "correlationId"
        in: "header"
        description: "A tracking id provided by the calling application"
        required: false
        type: "string"
      responses:
        "200":
          description: "Request was successful\n"
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/masterList"
        "204":
          description: "No Content Found."
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
        "400":
          description: "Bad request."
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
        "415":
          description: "Unsupported Content Type"
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
        "500":
          description: "Internal server error."
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
        "501":
          description: "Service under development - Not Implemented"
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
        "503":
          description: "Service Unavailable."
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
        "504":
          description: "Service Timeout Error"
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
      x-swagger-router-controller: "SoftwareList"
  /updateMasterList:
    post:
      tags:
      - "Software Master List"
      summary: "Update software list"
      description: "Gets entire monitored software list."
      operationId: "updateSoftwareList"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "correlationId"
        in: "header"
        description: "A tracking id provided by the calling application"
        required: false
        type: "string"
      responses:
        "200":
          description: "Request was successful\n"
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/masterList"
        "204":
          description: "No Content Found."
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
        "400":
          description: "Bad request."
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
        "415":
          description: "Unsupported Content Type"
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
        "500":
          description: "Internal server error."
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
        "501":
          description: "Service under development - Not Implemented"
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
        "503":
          description: "Service Unavailable."
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
        "504":
          description: "Service Timeout Error"
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
      x-swagger-router-controller: "SoftwareList"
  /deleteMasterList:
    delete:
      tags:
      - "Software Master List"
      summary: "Delete record in software list"
      description: "Gets entire monitored software list."
      operationId: "deleteSoftwareList"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "correlationId"
        in: "header"
        description: "A tracking id provided by the calling application"
        required: false
        type: "string"
      responses:
        "200":
          description: "Request was successful\n"
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/masterList"
        "204":
          description: "No Content Found."
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
        "400":
          description: "Bad request."
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
        "415":
          description: "Unsupported Content Type"
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
        "500":
          description: "Internal server error."
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
        "501":
          description: "Service under development - Not Implemented"
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
        "503":
          description: "Service Unavailable."
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
        "504":
          description: "Service Timeout Error"
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
      x-swagger-router-controller: "SoftwareList"
  /getUserList:
    get:
      tags:
      - "User List"
      summary: "Get user list and user software list based on admin id"
      operationId: "getUserList"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "correlationId"
        in: "header"
        description: "A tracking id provided by the calling application"
        required: false
        type: "string"
      responses:
        "200":
          description: "Request was successful\n"
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/userList"
        "204":
          description: "No Content Found."
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
        "400":
          description: "Bad request."
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
        "415":
          description: "Unsupported Content Type"
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
        "500":
          description: "Internal server error."
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
        "501":
          description: "Service under development - Not Implemented"
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
        "503":
          description: "Service Unavailable."
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
        "504":
          description: "Service Timeout Error"
          headers:
            correlationId:
              type: "string"
              description: "A tracking id provided by calling application"
            transactionId:
              type: "string"
              description: "Return the unique id of request generated by services"
          schema:
            $ref: "#/definitions/errorDefinition"
      x-swagger-router-controller: "SoftwareList"
securityDefinitions:
  Bearer:
    description: "Authorization token to verify the user accessing the endpoint"
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
  masterList:
    type: "object"
    properties:
      ready:
        type: "boolean"
        example: true
        description: "Status indicating health of backend micro-services used by this\
          \ API, true only if all backends are ready."
        default: false
      readyAsOf:
        type: "string"
        format: "date-time"
        example: "2018-08-07t15:12:22z"
        description: "Current date and time YYYY-MM-DDtHH:mm:ssz"
    example:
      ready: true
      readyAsOf: "2018-08-07t15:12:22z"
  errorDefinition:
    type: "object"
    properties:
      name:
        type: "string"
        example: "AuthenticationError"
        description: "Predefined name for the error"
      code:
        type: "integer"
        example: 50004
      message:
        type: "string"
        description: "Error message"
      fields:
        type: "array"
        items:
          $ref: "#/definitions/errorDefinition_fields"
    example:
      name: "AuthenticationError"
      code: 50004
      message: "Error"
      fields:
      - name: "fieldId1"
        value: "Incorrect format"
      - name: "fieldId2"
        value: "Incorrect format"
  errorDefinition_fields:
    properties:
      name:
        type: "string"
        example: "fieldId"
        description: "Error field name"
      value:
        type: "string"
        example: "Incorrect format"
        description: "Issue with the fieldId, if any."
  userList:
    type: "object"
    properties:
      clientBalance:
        type: "array"
        description: "Array of balance records for given account number"
        items:
          $ref: "#/definitions/records"
    example:
      clientBalance:
        records:
        - accountNumber: "634346334"
          balance: 3453434
          asOfDate: "2018-03-14"
          dayChangeAmt: 234232
          dayChangePct: 0.1000000000000000055511151231257827021181583404541015625
          accruedInterest: 192984
  records:
    type: "object"
    properties:
      accountNumber:
        type: "string"
        description: "Account Number"
      balance:
        type: "number"
        description: "Balance"
      asOfDate:
        type: "string"
        format: "date"
        description: "Record is latest as of this date"
      pendingDividend:
        type: "string"
        description: "Pending Dividend"
      marginBalance:
        type: "string"
        description: "Margin Balance"
      dayChangeAmt:
        type: "number"
        description: "Amount of Day Change"
      dayChangePct:
        type: "number"
        description: "Percentage of Day Change"
      accruedInterest:
        type: "number"
        description: "Accrued Interest"
    example:
      accountNumber: "634346334"
      balance: 325235223
      asOfDate: "2018-03-14"
      pendingDividend: "2423422"
      marginBalance: "242352"
      dayChangeAmt: 234232
      dayChangePct: 2323111
      accruedInterest: 192984